feat(harness): make kanban the home page and drop Jobs dashboard
Why

The harness landing page was still a period-stats strip plus a collapsible
Jobs card (Working / Failed / Completed). Operators live on the pipeline
board; that home surface was no longer the right default. Empty instances
also never showed the add-repo blank slate unless you opened `/repos`.

What changed

- `/` is the kanban board when at least one repository is configured
  (board UI extracted to `kanban-board.tsx`).
- Zero repositories: `/` reuses the shared add-repo blank slate
  (`EmptyRepositoriesBlankSlate`); no empty pipeline.
- Nav: remove the separate Home tab; product title and Kanban both
  target `/`. Order: Repos, Kanban, Completed.
- `/kanban` redirects to `/` with `replace: true` so bookmarks and deep
  links land on home without re-entering the legacy path on Back.
- Remove home Jobs UI (Working / Failed / Completed card and related
  dead tests); committed-PR totals stay above the board.
- Live updates: mount repository membership SSE on home so CLI
  add/remove refreshes the blank-slate to board gate without navigating
  away.
- Soft-fail repositories load on home (loading / error+retry / blank /
  board) instead of suspense-only unmount.
- Docs and e2e updated for the new home URL, nav, and empty-state
  behavior.

Verification

- `bunx nx test harness` (unit suite green after the change set)
- `bunx nx build harness`
- Pre-commit (Biome/lint pipeline) passed after staging format fixes
- Review remediations: membership SSE and soft-fail gate; docs wording;
  `replace: true` redirect

Limitations / out of scope

- Kanban column/lane behavior unchanged.
- `/repos` still has its own add-repo UI when repositories already exist.
- Full-bleed width still applies to the entire `/` path (including blank
  slate); cosmetic only vs `/repos` empty state.

Closes #684
