+++ apps/harness/test/primary-nav.test.ts
@@ -8,18 +8,26 @@ const rootSource = () =>
 const stylesSource = () =>
   readFileSync(join(import.meta.dir, "../src/styles.css"), "utf8")
 
-describe("primary Home / Repos / Kanban / Completed navigation", () => {
-  test("root chrome exposes Home, Repos, Kanban, and Completed Link controls", () => {
+describe("primary Repos / Kanban / Completed navigation", () => {
+  test("root chrome exposes Repos, Kanban, and Completed Link controls without Home", () => {
     const source = rootSource()
     expect(source).toContain('aria-label="Primary"')
     expect(source).toContain('to="/"')
     expect(source).toContain('to="/repos"')
-    expect(source).toContain('to="/kanban"')
     expect(source).toContain('to="/completed"')
-    expect(source).toMatch(/Home\s*<\/Link>/)
     expect(source).toMatch(/Repos\s*<\/Link>/)
     expect(source).toMatch(/Kanban\s*<\/Link>/)
     expect(source).toMatch(/Completed\s*<\/Link>/)
+    // Distinct Home tab is gone; logo and Kanban both target `/`.
+    expect(source).not.toMatch(/Home\s*<\/Link>/)
+    expect(source).not.toContain("function HomeNavIcon()")
+    expect(source).not.toContain("<HomeNavIcon />")
+    // Kanban lives at home — no separate /kanban nav target.
+    const settingsBlock = source.slice(
+      source.indexOf("<SettingsButton"),
+      source.indexOf("</nav>"),
+    )
+    expect(settingsBlock).not.toContain('to="/kanban"')
   })
 
   test("primary nav uses a full-width bold black divider on every route", () => {
@@ -49,7 +57,7 @@ describe("primary Home / Repos / Kanban / Completed navigation", () => {
     )
   })
 
-  test("groups Home, Repos, Kanban, Completed, and Settings in one right-aligned control cluster", () => {
+  test("groups Repos, Kanban, Completed, and Settings in one right-aligned control cluster", () => {
     const source = rootSource()
     const clusterMarker =
       'className="ml-auto flex items-center gap-2 self-center"'
@@ -65,20 +73,16 @@ describe("primary Home / Repos / Kanban / Completed navigation", () => {
       source.indexOf("</nav>"),
     )
     expect(settingsBlock).toContain("leading={")
-    expect(settingsBlock).toContain('to="/"')
     expect(settingsBlock).toContain('to="/repos"')
-    expect(settingsBlock).toContain('to="/kanban"')
+    expect(settingsBlock).toContain('to="/"')
     expect(settingsBlock).toContain('to="/completed"')
-    expect(settingsBlock).toContain("<HomeNavIcon />")
     expect(settingsBlock).toContain("<ReposNavIcon />")
     expect(settingsBlock).toContain("<KanbanNavIcon />")
     expect(settingsBlock).toContain("<CompletedNavIcon />")
-    const homeIdx = settingsBlock.indexOf('to="/"')
     const reposIdx = settingsBlock.indexOf('to="/repos"')
-    const kanbanIdx = settingsBlock.indexOf('to="/kanban"')
+    const kanbanIdx = settingsBlock.indexOf('to="/"')
     const completedIdx = settingsBlock.indexOf('to="/completed"')
-    expect(homeIdx).toBeGreaterThan(-1)
-    expect(reposIdx).toBeGreaterThan(homeIdx)
+    expect(reposIdx).toBeGreaterThan(-1)
     expect(kanbanIdx).toBeGreaterThan(reposIdx)
     // Completed sits after Kanban (and after Repos).
     expect(completedIdx).toBeGreaterThan(kanbanIdx)
@@ -102,19 +106,16 @@ describe("primary Home / Repos / Kanban / Completed navigation", () => {
     expect(brandIdx).toBeLessThan(source.indexOf("<SettingsButton"))
   })
 
-  test("Home, Repos, Kanban, and Completed use stroke icons matching Settings icon language", () => {
+  test("Repos, Kanban, and Completed use stroke icons matching Settings icon language", () => {
     const source = rootSource()
-    expect(source).toContain("function HomeNavIcon()")
     expect(source).toContain("function ReposNavIcon()")
     expect(source).toContain("function KanbanNavIcon()")
     expect(source).toContain("function CompletedNavIcon()")
-    expect(source).toContain("<HomeNavIcon />")
     expect(source).toContain("<ReposNavIcon />")
     expect(source).toContain("<KanbanNavIcon />")
     expect(source).toContain("<CompletedNavIcon />")
     // Icons: aria-hidden, size-3.5, stroke currentColor (same as Settings gear).
     for (const iconFn of [
-      "HomeNavIcon",
       "ReposNavIcon",
       "KanbanNavIcon",
       "CompletedNavIcon",
@@ -142,14 +143,15 @@ describe("primary Home / Repos / Kanban / Completed navigation", () => {
     expect(source).toContain(
       "activeProps={{ className: primaryNavLinkActiveClassName }}",
     )
-    // Home nav control (not the brand title Link) must use exact matching.
-    // Slice from the SettingsButton leading prop through the first non-Home destination.
-    const homeSwitcherLink = source.slice(
-      source.indexOf("leading={"),
+    // Kanban nav control (not the brand title Link) must use exact matching.
+    // Slice from the Repos link through Completed so the Kanban `to="/"` is included.
+    const kanbanSwitcherLink = source.slice(
       source.indexOf('to="/repos"'),
+      source.indexOf('to="/completed"'),
     )
-    expect(homeSwitcherLink).toContain('to="/"')
-    expect(homeSwitcherLink).toContain("activeOptions={{ exact: true }}")
+    expect(kanbanSwitcherLink).toContain('to="/"')
+    expect(kanbanSwitcherLink).toContain("<KanbanNavIcon />")
+    expect(kanbanSwitcherLink).toContain("activeOptions={{ exact: true }}")
     // Shared base must not carry exclusive active/inactive visual tokens —
     // Router merges className with active/inactive props.
     const sharedClassDecl = source.match(
@@ -192,26 +194,24 @@ describe("primary Home / Repos / Kanban / Completed navigation", () => {
       /const primaryNavActionClassName =\s*\n?\s*"[^"]*text-ink-faint[^"]*"/,
     )
     // Destinations are client Links, not raw <a href> only.
-    expect(source).not.toMatch(/<a\s+href=["']\/kanban["']/)
     expect(source).not.toMatch(/<a\s+href=["']\/repos["']/)
     expect(source).not.toMatch(/<a\s+href=["']\/completed["']/)
     expect(source).not.toMatch(/<a\s+href=["']\/["']/)
   })
 
-  test("shared nav lives in root layout so Home, Repos, Kanban, and Completed inherit it", () => {
+  test("shared nav lives in root layout so Repos, Kanban, and Completed inherit it", () => {
     const source = rootSource()
     const rootComponent = source.slice(
       source.indexOf("function RootComponent("),
     )
     expect(rootComponent).toContain('to="/repos"')
-    expect(rootComponent).toContain('to="/kanban"')
     expect(rootComponent).toContain('to="/completed"')
-    expect(rootComponent).toMatch(/Home\s*<\/Link>/)
     expect(rootComponent).toMatch(/Repos\s*<\/Link>/)
     expect(rootComponent).toMatch(/Kanban\s*<\/Link>/)
     expect(rootComponent).toMatch(/Completed\s*<\/Link>/)
+    expect(rootComponent).not.toMatch(/Home\s*<\/Link>/)
     expect(rootComponent).toContain("<Outlet />")
-    expect(rootComponent.indexOf("Home")).toBeLessThan(
+    expect(rootComponent.indexOf("Repos")).toBeLessThan(
       rootComponent.indexOf("<Outlet />"),
     )
   })
